refactor: shared, entities, widgets 분리
