+++ src/__mocks__/handlers.ts
@@ -1,7 +1,7 @@
 import { http, HttpResponse } from 'msw';
 
 import { events } from '../__mocks__/response/events.json' assert { type: 'json' };
-import { Event } from '../types';
+import { Event } from '../entities/event/model/types';
 
 export const handlers = [
   http.get('/api/events', () => {
