+++ src/__tests__/unit/easy.timeValidation.spec.ts
@@ -1,4 +1,4 @@
-import { getTimeErrorMessage } from '../../utils/timeValidation';
+import { getTimeErrorMessage } from '../../shared/lib/timeValidation';
 
 describe('getTimeErrorMessage >', () => {
   it('시작 시간이 종료 시간보다 늦을 때 에러 메시지를 반환한다', () => {
