+++ src/entities/calendar/config/constants.ts
@@ -1,0 +1,3 @@
+export const CELL_WIDTH = '14.28%';
+export const CELL_HEIGHT = '100px';
+export const weekDays = ['일', '월', '화', '수', '목', '금', '토'];
