+++ src/entities/calendar/model/types.ts
@@ -1,0 +1,6 @@
+export interface CalendarViewProps {
+  currentDate: Date;
+  events: Event[];
+  notifiedEvents: string[];
+  holidays?: Record<string, string>;
+}
