+++ src/entities/event/lib/eventUtils.ts
@@ -1,5 +1,5 @@
-import { Event } from '../types';
-import { getWeekDates, isDateInRange } from './dateUtils';
+import { getWeekDates, isDateInRange } from '../../calendar/lib/dateUtils';
+import { Event } from '../model/types';
 
 /**
  * 특정 날짜 범위 내의 이벤트들만 필터링합니다.
