+++ src/entities/event/model/types.ts
