+++ src/shared/config/index.ts
@@ -1,0 +1,3 @@
+export { categories } from './categories.ts';
+export { notificationOptions } from './notificationOptions.ts';
+export { weekDays } from './weekDays.ts';
